Made it possible to track threads based on a ticket number in the subject.
TicketNumber module required and I also have a small update for this one.
Added .env settings for:
TICKETNUMBER_PREFIX defaulting to "[#"
TICKETNUMBER_SUfFIX defaulting to "] "
TICKETNUMBER_TRUSTSUBJECT defaulting to false (this will toggle this functionality on and odd)
This fix will be great for everyone who need to forward tickets to external helpdesks that are not always adhering to standard use of In-Reply-To and Referenced headers.

If the subject matches a conversation_id it will use the first thread in that conversation as its parent messageid