Add header indicating mailbox name
By adding a header that references the mailbox name, it is easier for support agents to set their desired filter rules. 

For example, an individual support agent may want notifications from some mailboxes, but not others. He turns on notifications on his account, but then has to use email filters to delete the ones he did not want. Or, he wishes to redirect notifications for different mailboxes to different final destinations.