Add APP_KEY_FILE environment variable
In lots environments having files with secrets is common (e.g. with
NixOS where static secrets in config files are not secure). This new
environment variable adds support for this. If a APP_KEY is defined it
will take precedence ofer APP_KEY_FILE
Tests have been added to show this beaviour.